backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.4
sqlalchemy==2.0.28
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.1
python-jose[cryptography]>=3.3.0
httpx<0.28.0
